requirements.txt
Flask==0.11.1
Flask-Script==2.0.5
beautifulsoup4==4.5.3
TestLink-API-Python-client==0.4.7
Flask-SQLAlchemy==2.1